backend/requirements.txt

# Backend API Dependencies
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic[email]

# Async Support
httpx==0.27.2
aiofiles==24.1.0

# Database
motor==3.3.2

# Email Validation
email-validator==2.2.0

# CORS & Security
python-multipart==0.0.12
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
